requirements.txt
# Python 3.10 用 core stack
numpy==2.2.5
pandas==2.2.3
python-dateutil==2.9.0.post0

# Google Cloud / Vertex AI
google-cloud-bigquery==3.31.0
vertexai==1.71.1
google-generativeai==0.8.5   # レガシー SDK（2025‑08‑31 EOL）

# ユーティリティ
tqdm==4.67.1
python-dotenv==1.1.0         # import 名は「dotenv」
beautifulsoup4==4.13.4

# アプリ/UI
streamlit==1.45.0
nest_asyncio==1.5.5
